Gubbins serial generator/validator
The Gubbins class and associated scripts generate serials as per, given a
Prefix and a 32-bit ID.

The Prefix is an arbitrary byte array and generates a Seed and an initial
Checksum.  The Seed is produced by feeding lower-case version of the Prefix
to SHAKE 128 and taking a 32-bit digest in big-endian order.  The Checksum
is the XOR of these Seed bytes in big-endian order.

The ID is XORd with the Seed, passed into triple32 [1], which is reversible
over the 2**32 field and distributed under the Unlicense.  The result XORd
again with the Seed to generate the Point ID.  This protects our ID from
obvious enumeration and causes the resulting Serial to depend on the Prefix.
Since the resulting bytes are big-endian, we ensure that appending values to
the end of the sequence modifies the resulting ID.  We then take a Checksum
over these bytes and XOR them with Prefix Checksum to generate the Final
Checksum; this is appeneded to the Point ID, and the concatenation is base32
encoded using a modified zbase32 [2] dictionary.  This encoding is split
into 4-character chunks with dashes as a separator.

The zbase32 dictionary has several desirable properties including that
characters do not appear sequentially and that more frequent points, e.g.
at the end of a sequence, are expressed by easier-to-read characters.

In zbase32, u is removed because it might be confused with u.  This means
that i and 1 are both valid and given an I, we don't know how to fix it up.
Thus, we replace i with u and can translate both l and i to 1.  To deal with
other potential confusion issues, we convert some of the characters to
upper-case but accept them as lower-case in the validator.

Our converted characters are IL=>1, O=>0, 2=>Z.  This is immutable.

Our upper-cased characters are currently AJNRTUV, trying to avoid lower-case
confusion respectively with oe, i, r, n, i, v, u.  This is mutable, since
the validator will always treat characters as lower-case.

To validate, reverse the steps above.  A canonicalize function is also
provided to fix up abritrary Serial Strings that are valid but illicitly
formated.

We also provide gubbins.generate and gubbins.validate as scripts that
generate and validate Serials.  Both are able to take arguments on either
the command line or as part of stdin.  The command line for gubbins is able
to take ranges via pynumparser, so the following should work as expected

	python -m gubbins.generate --validate --solo --prefix=AX 1-10 20-21 |
	python -m gubbins.validate --validate --solo

and test the underlying system.

It is stupid easy to enumerate IDs given an existing ID and Prefix.  To
avoid that, we would only want to populate some of the valid space.

[1] https://github.com/skeeto/hash-prospector
[2] https://github.com/alanblevins/anybase32