Translate ambiguous prefix characters
To avoid someone getting confused about 012 in the prefix, add translations
from oO to 0, iIlL to 1, and zZ to 2 in the prefix.  When generating new
serials with the generator tool, assert that the prefix is canonical.
Modify the canonicalize method to validate and then re-generate the serial
from scratch, rather than just doing character translation.