Add support for Additional Data
When using serials to assert ownership of devices, it is important to avoid
enumeration attacks.  Since a gubbins serial can be converted back to a ID,
this means either

	* avoiding sequences of consecutive IDs, or
	* mixing up the sequence of IDs that come out of the generator.

The former is possible but the CLI generator allows sequential generation
and bookkeeping is an exercise for the caller.  Therefore, allow callers to
provide additional data that mixes the sequence but is not part of the
serial itself.  When validating a serial produced with AD, the validator
MUST know the AD out-of-band.

Note that ad=None and ad="" produce different results.